refactor: 고객지원 페이지 서버 데이터 조회 로직 분리
